+++ src/kioskboard.d.ts
@@ -1,0 +1,3 @@
+import KioskBoard from '../index.d';
+export = KioskBoard;
+export as namespace KioskBoard;